stack: Fix display of size in interactive stack
On physical calculators, the `%zu` format for `printf` is no longer
supported after switching to the nano library. This caused the display
of information in the interactive stack to be wrong, showing `zu`
instread instead of the size, and then showing bogus information for
the type.

Replace `%zu` with `%lu` and convert argument to the `ulong` time.

+++ src/user_interface.cc
@@ -4621,8 +4621,8 @@ bool user_interface::handle_editing(int key)
                     {
                         char sizebuf[40];
                         char valbuf[40];
-                        snprintf(sizebuf, sizeof(sizebuf), "%zu bytes %s",
-                                 size, obj->fancy());
+                        snprintf(sizebuf, sizeof(sizebuf), "%lu bytes %s",
+                                 (unsigned long) size, obj->fancy());
                         bin->render(valbuf, sizeof(valbuf));
                         draw_message("Object info", sizebuf, valbuf);
                         wait_for_key_press();
